% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aSPU.R
\name{aSPU}
\alias{aSPU}
\title{Weighted adaptive sum of powered score tests (waSPU) test for single trait with single or multiple weights for each genetic marker}
\usage{
aSPU(U, V, weight, pow = c(1:8, Inf), n.perm = 1000)
}
\arguments{
\item{U}{Score vector for the genetic marker set (p by 1 matrix).}

\item{V}{Corresponding covariance matrix for the score vector (p by p matrix).}

\item{weight}{Single weight for each genetic markers (p by 1 matrix).}

\item{pow}{Gamma sets. Different gamma corresponds to different existing test. For example, gamma = 1 equals Sum test; gamma = 2 equals SSU or SKAT with linear kernel; gamma = Inf is similar to minimum p value test. We recommend use pow = c(1:6, Inf) to maintain high power under various scenarios, although other choice may be slightly powerful in some specific situations.}

\item{n.perm}{number of permutations.}
}
\value{
P-values for aSPU test.
}
\description{
It returns p-values.
}
\author{
Chong Wu, Zhiyuan Xu, and Wei Pan
}

