% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aSPUO.R
\name{aSPUO}
\alias{aSPUO}
\title{Adaptive Sum of Powered Score Omnibus tests (aSPUO) test for single trait with multiple weights for each genetic marker}
\usage{
aSPUO(U, V, weight, pow = c(1:8, Inf), n.perm = 1000)
}
\arguments{
\item{U}{Score vector for the genetic marker set (p by 1 matrix).}

\item{V}{Corresponding covariance matrix for the score vector (p by p matrix).}

\item{weight}{Multiple weights for each genetic markers (p by p matrix).}

\item{pow}{Gamma sets. Different gamma corresponds to different existing test. For example, gamma = 1 equals Sum test; gamma = 2 equals SSU or SKAT with linear kernel; gamma = Inf is similar to minimum p value test. We recommend use pow = c(1:6, Inf) to maintain high power under various scenarios, although other choice may be slightly powerful in some specific situations.}

\item{n.perm}{number of permutations.}
}
\value{
It returns a list, which contains the following two information:
         asy-SPU(1)-O: Theoretical p value for SPU(1)-O. This p value is exactly same as TWAS-Omnibus.
         sim-pvalue: a matrix containing p values for SPU/aSPU with different weights, and p values for SPU-O and aSPU-O.
}
\description{
It returns p-values.
}
\author{
Chong Wu, Zhiyuan Xu, and Wei Pan
}

