% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aSPUPath.R
\name{aSPUPath}
\alias{aSPUPath}
\title{Pathway based Sum of Powered Score tests (SPUsPath) and adaptive SPUpath (aSPUsPath) test for single trait - pathway association starting with score vector and its covariance matrix.}
\usage{
aSPUPath(U, V, info, pow = c(1, 2, 4, 8, Inf), pow2 = c(1, 2, 4, 8),
  n.perm = 1000)
}
\arguments{
\item{U}{Score vector for the marker set we are interested in.}

\item{V}{Corresponding covariance matrix for the score vector.}

\item{info}{A table for mapping each marker to the specific genes and chromosome.}

\item{pow}{SNP or single CpG sites specific power(gamma values) used in SPUpath test.}

\item{pow2}{GENE specific power(gamma values) used in SPUpath test.}

\item{n.perm}{number of permutations.}
}
\value{
P-values for SPUMpath tests and aSPUMpath test.
}
\description{
It gives p-values of the SPUsPath tests and aSPUsPath test starting with score vector and its covariance matrix.
}
\references{
Il-Youp Kwak, Wei Pan (2015)
Adaptive Gene- and Pathway-Trait Association Testing with GWAS Summary Statistics,
Bioinformatics, 32(8):1178-1184
}
\author{
Chong Wu and Wei Pan
}
