% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daSPU.R
\name{daSPU}
\alias{daSPU}
\title{Doubley Adaptive Sum of Powered Score tests (daSPU) test for single trait with multiple weights for each genetic marker}
\usage{
daSPU(U, V, weight, pow1 = c(1:8, Inf), pow2 = c(1:8, Inf), n.perm = 1000)
}
\arguments{
\item{U}{Score vector for the marker set we are interested in. (N by 1 matrix)}

\item{V}{Corresponding covariance matrix for the score vector. (N by N matrix)}

\item{weight}{Multiple weights for each genetic markers. (N by M matrix)}

\item{pow1}{SNP or single CpG sites specific power(gamma values) used in daSPU test.}

\item{pow2}{Specific power(gamma values) used for different weight.}

\item{n.perm}{number of permutations.}
}
\value{
P-values for daSPU test.
}
\description{
It returns p-values and test statistics
}
\author{
Chong Wu and Wei Pan
}

