% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_score.R
\name{glmm_score}
\alias{glmm_score}
\title{GLMM Score}
\usage{
glmm_score(fit0, G)
}
\arguments{
\item{fit0}{A generalized linear mixed model fitted under the null hypothesis using the "glmmkin" function of the GMMAT R package, where p is the number of variants.}

\item{G}{A n * p genotype matrix. n is the number of samples and the order must correspond with fit0.}
}
\value{
score vector (U) and corresponding covariance matrix (V).
}
\description{
It returns the score vector and its corresponding covariance matrix.
}
\author{
Chong Wu, Jun Young Park, and Wei Pan
}
